+++ src/main/java/net/brdle/collectorsreap/data/CRItemTags.java
@@ -48,6 +48,9 @@ public class CRItemTags {
 	public static final TagKey<Item> RAW_STRIDER = forge("raw_strider");
 	public static final TagKey<Item> COOKIES = forge("cookies");
 	public static final TagKey<Item> CHEESE = forge("cheese");
+	public static final TagKey<Item> CHEESE_MILD = forge("cheese/mild");
+	public static final TagKey<Item> CHEESE_SPICY = forge("cheese/spicy");
+	public static final TagKey<Item> CHEESE_SWEET = forge("cheese/sweet");
 	public static final TagKey<Item> ROE = forge("roe");
 	public static final TagKey<Item> ROE_FISH = forge("roe/fish");
 	public static final TagKey<Item> ROE_PRAWN = forge("roe/prawn");