+++ changelog.md
@@ -21,6 +21,7 @@
 - Added Hermit's Sour (for [Brewin' and Chewin'](https://www.curseforge.com/minecraft/mc-mods/brewin-and-chewin))
 - Added Rosé Moon (for [Brewin' and Chewin'](https://www.curseforge.com/minecraft/mc-mods/brewin-and-chewin))
 - Added Reanimator's Garden (for [Brewin' and Chewin'](https://www.curseforge.com/minecraft/mc-mods/brewin-and-chewin) and [Atmospheric](https://www.curseforge.com/minecraft/mc-mods/atmospheric))
+- Added Cream Cheese
 - Lime Green Tea is now brewed from Green Tea with [Farmer's Respite](https://www.curseforge.com/minecraft/mc-mods/farmers-respite)
 - Pomegranate Black Tea is now brewed from Black Tea with [Farmer's Respite](https://www.curseforge.com/minecraft/mc-mods/farmers-respite)
 - Clam Chowder now provides Comfort for 3:00